test: reconcile unified gestures with helper ownership
